Moved live submission to its own component
