+++ .gitignore
@@ -5,3 +5,4 @@ build
 node_modules
 ncp-debug.log
 npm-debug.log
+src/config.js
