--- src/constants/ActionTypes.js
@@ -1,14 +1,0 @@
-/**
- * React Starter Kit (https://www.reactstarterkit.com/)
- *
- * Copyright © 2014-2016 Kriasoft, LLC. All rights reserved.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE.txt file in the root directory of this source tree.
- */
-
-import keyMirror from 'fbjs/lib/keyMirror';
-
-export default keyMirror({
-
-});
