+++ src/reducers/liveSubmission.js
@@ -1,4 +1,4 @@
-import { SET_SUBMISSION_SHOWN, SET_SUBMISSION_LIMIT } from '../actions/headerAction';
+import { SET_SUBMISSION_SHOWN, SET_SUBMISSION_LIMIT } from '../actions/liveSubmissionAction';
 
 const defaultState = {
   submission: {
@@ -9,7 +9,7 @@ const defaultState = {
 };
 
 
-const header = (state = defaultState, action) => {
+const liveSubmission = (state = defaultState, action) => {
   switch (action.type) {
     case SET_SUBMISSION_SHOWN:
       return {
@@ -32,4 +32,4 @@ const header = (state = defaultState, action) => {
   }
 };
 
-export default header;
+export default liveSubmission;
